requirements.txt
Flask==2.3.3
numpy==1.24.3
qrcode[pil]==7.4.2
Pillow==10.0.0